Refactor: Update CUDA kernel computations to use `float` instead of `double`
- Changed data types in CUDA mixed GPU implementation from `double` to `float` for improved performance and lower memory usage.
- Added compile-time guards in `test_mixed_gpu` to handle scenarios with or without CUDA support (`USE_CUDA`).

+++ src/cuda/mixed_gpu.h
@@ -32,7 +32,7 @@ typedef struct {
 } KernelPool;
 
 typedef struct {
-    double *kernel_pool;
+    float *kernel_pool;
     int kernel_pool_size;
 
     int *kernel_offsets;
